% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microData}
\alias{microData}
\title{Micro data}
\format{
\subsection{\code{microData}}{

A data frame with 20 rows and 10 columns:
}
}
\usage{
microData
}
\description{
A sample of data
Report ...
}
\keyword{datasets}
