% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microSummary}
\alias{microSummary}
\title{Micro summary data}
\format{
\subsection{\code{microSummary}}{

A data frame with 111 rows and 13 columns:
}
}
\usage{
microSummary
}
\description{
A sample of data
Report ...
}
\keyword{datasets}
