\name{MSToolkit-package}
\alias{MSToolkit-package}
\alias{MSToolkit}
\docType{package}
\title{
Generates and analyzes simulated trial data
}
\description{
Originally designed and built for Pfizer by Mango Solutions, the MSToolkit library generates and analyzes simulated trial data.
}
\details{
\tabular{ll}{
Package: \tab MSToolkit\cr
Type: \tab Package\cr
Version: \tab 1.6\cr
Date: \tab 2007-08-29\cr
}}

\author{ Mango Solutions & Pfizer \email{mstoolkit@googlemail.com} } 

\keyword{ package }
\seealso{
  Neal Thomas' "DoseResponse" library
}
\examples{

}
