% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePredCode.R
\name{parsePredCode}
\alias{parsePredCode}
\title{parsePredCode}
\usage{
parsePredCode(model, respCol = getEctdColName("Response"))
}
\arguments{
\item{model}{...}

\item{respCol}{...}
}
\description{
Converts NONMEM Modelling code to executable S code, returning either a list of commands or a vector of commands
}
\author{
Rich
}
