% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ReplicateSample}
\alias{ReplicateSample}
\title{Replicate Sample data}
\format{
\subsection{\code{ReplicateSample}}{

A data frame with 500 rows and 14 columns:
}
}
\usage{
ReplicateSample
}
\description{
A sample of data
Report ...
}
\keyword{datasets}
