\name{tvDefinition-class}
\docType{class}
\alias{tvDefinition-class}


\title{tvDefinition class}
\description{tvDefinition class}

\section{Slots}{
 \describe{
    \item{\code{Name}:}{Name }
    \item{\code{Value}:}{Value}
    \item{\code{trtCall}:}{trtCall}
  }
}
\author{ Mango Solutions & Pfizer \email{mstoolkit@googlemail.com}  }
\keyword{classes}
