% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkColNames.R
\name{checkColNames}
\alias{checkColNames}
\title{Check that column names are found in a data frame}
\usage{
checkColNames(dNames, cols)
}
\arguments{
\item{dNames}{Character vector of column names from a data frame}

\item{cols}{Character vector of column names that should be found in the
data frame}
}
\value{
A TRUE value (if all of "cols" are found in "dNames") or an
exception if not
}
\description{
Checks that all column names specified in "cols" are found in the data frame
names (given by "dNames") and produces a neat error message if not
}
\examples{

	checkColNames(LETTERS, "D")
	checkColNames(LETTERS, LETTERS[1:5])
	\dontrun{
	  # Will show an error because "Hello" is not in LETTERS
	  checkColNames(LETTERS, c("A", "Hello"))
	}

}
\author{
Romain Francois
}
