% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMacroFormat.R
\name{checkMacroFormat}
\alias{checkMacroFormat}
\title{Check the macro format dataset}
\usage{
checkMacroFormat(data)
}
\arguments{
\item{data}{(Required) Dataset to be checked to ensure it is a valid "Macro
Evaluation" structure}
}
\value{
Nothing. Only used for error generation side effects.
}
\description{
This function is used to perform checks on the structure of the macro format
dataset produced by the user-supplied macro code in
}
\examples{

  checkMacroFormat(iris[1,])

  \dontrun{
    # not a data frame
    checkMacroFormat( 1:10 )

    # more than one row
    checkMacroFormat(iris[1:2,])

  }

}
\author{
: Francisco Gochez, Rich Pugh
}
\keyword{IO}
