% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSymmetricPDMatrix.R
\name{checkSymmetricPDMatrix}
\alias{checkSymmetricPDMatrix}
\title{Checks if a matrix is symmetric and positive definite}
\usage{
checkSymmetricPDMatrix(mat, tol = 1e-06)
}
\arguments{
\item{mat}{(Required) The matrix to check}

\item{tol}{(Optional) The tolorence to use when comparing the value to 0.
The default is "1e-06"}
}
\value{
Returns an error if the matrix does not satisfy symmetric or
positive definite. Otherwise does not return anything
}
\description{
Checks if a matrix is symmetric and positive definite using an eigen value
decomposition.
}
\examples{
  checkSymmetricPDMatrix( diag(4))
  A <- matrix( c( 2, -1, 0, -1, 2, -1, 0, -1, 2 ), nrow=3, byrow=TRUE )
  checkSymmetricPDMatrix( A )
  \dontrun{
  # Not symmetric
  B <- matrix( c( 1, 2, 3, 4, 5, 6, 7, 8, 9), nrow=3, byrow=TRUE )
  checkSymmetricPDMatrix(B)
  # Not positive definite
  C <- matrix( c( -2, 1, 0, 1, -2, 1, 0, 1, -2 ), nrow=3, byrow=TRUE )
  checkSymmetricPDMatrix(C)
  }
}
\author{
Romain Francois
}
\keyword{error}
