% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSASIncludeRows.R
\name{convertSASIncludeRows}
\alias{convertSASIncludeRows}
\title{Create a SAS 'IF' statement}
\usage{
convertSASIncludeRows(
  includeRows = NULL,
  doseCol = getEctdColName("Dose"),
  interimCol = getEctdColName("Interim")
)
}
\arguments{
\item{includeRows}{Row inclusion matrix (2 columns: Interim value and Dose
value)}

\item{doseCol}{Dose column name, given by \link{getEctdColName} by default}

\item{interimCol}{Interim column name, given by \link{getEctdColName} by
default}
}
\value{
A single character string
}
\description{
Create a SAS 'IF' statement, used to describe rows to include in a SAS
analysis
}
\details{
Takes a matrix of interim and dose values, and produces a "SAS style" IF
statement
}
\examples{

convertSASIncludeRows( cbind(rep(1:2, each = 3), c(0, 15, 30, 0, 30, 45) ) )

}
\author{
Mike K Smith \email{mstoolkit@googlemail.com}
}
\keyword{SAS}
