% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParameters.R
\name{createParameters}
\alias{createParameters}
\title{Creates parameters for subjects in each replicate}
\usage{
createParameters(
  subjects,
  genNames,
  genFixedMean,
  genFixedCov = 0,
  genRange,
  genBetweenNames,
  genBetweenMean,
  genBetweenCov,
  genErrStruc = "None",
  genMaxDraws = 10,
  genParRangeTolerance = 0.5,
  extFile,
  extNames,
  extBetween,
  extBetweenNums,
  extSubset,
  extRange,
  extErrStruc = "None",
  extRefCol,
  extRefColName,
  extRefColSuffix,
  extIndEffects,
  extDataId = idCol,
  workingPath = getwd(),
  suffix = ".Between",
  idCol = getEctdColName("Subject"),
  seed = .deriveFromMasterSeed(),
  flagName = getEctdColName("ParOmit")
)
}
\description{
Creates simulated fixed and between subject parameters for subjects in each replicate
}
