% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParameters.R
\name{createParameters}
\alias{createParameters}
\title{Creates parameters for subjects in each replicate}
\usage{
createParameters(
  subjects,
  genNames,
  genFixedMean,
  genFixedCov = 0,
  genRange,
  genBetweenNames,
  genBetweenMean,
  genBetweenCov,
  genErrStruc = "None",
  genMaxDraws = 10,
  genParRangeTolerance = 0.5,
  extFile,
  extNames,
  extBetween,
  extBetweenNums,
  extSubset,
  extRange,
  extErrStruc = "None",
  extRefCol,
  extRefColName,
  extRefColSuffix,
  extIndEffects,
  extDataId = idCol,
  workingPath = getwd(),
  suffix = ".Between",
  idCol = getEctdColName("Subject"),
  seed = .deriveFromMasterSeed(),
  flagName = getEctdColName("ParOmit")
)
}
\arguments{
\item{subjects}{Subjects for which to create parameters}

\item{genNames}{Names of fixed effects to generate}

\item{genFixedMean}{Means for generating fixed parameters}

\item{genFixedCov}{Covariance Matrix for generating fixed parameter. \emph{0} by default.}

\item{genRange}{Range of Acceptable values for derived parameters}

\item{genBetweenNames}{Between subjects effects to generate}

\item{genBetweenMean}{Means for generated between subject effects}

\item{genBetweenCov}{Between subject effects covariance.}

\item{genErrStruc}{The function to map effects: \emph{Additive}, \emph{Proportional}, \emph{Log-Normal} or \emph{None}. The default is \emph{"None"}.}

\item{genMaxDraws}{The maximum number of draws. The default value is \emph{10}.}

\item{genParRangeTolerance}{The proportion of subjects with "in range" parameters with. The default value is \emph{0.5}.}

\item{extFile}{The external file name for data to import.}

\item{extNames}{Names of parameters to import (refers to variables in the input data)}

\item{extBetween}{Between subject effects variables in the data}

\item{extBetweenNums}{The integer mapping between random and fixed effects.}

\item{extSubset}{The subset to be applied to data before sampling.}

\item{extRange}{The range of Acceptable values for derived parameters.}

\item{extErrStruc}{Function to map between subject effects: \emph{Additive}, \emph{Proportional}, \emph{Log-Normal} or \emph{None}.
The default is \emph{"None"}.}

\item{extRefCol}{The column of reference data.}

\item{extRefColName}{The column name in data for referenced sampling.}

\item{extRefColSuffix}{Suffix to add to reference variable in the data.}

\item{extIndEffects}{Individual effects flag}

\item{extDataId}{The external subject variable name. By default,
this is taken to be the same as the "idCol" input.}

\item{workingPath}{Working path from which to import covariate file. By default, the working path will be used}

\item{suffix}{Suffix for retained between subject effects. The default is \emph{".Between"}.}

\item{idCol}{Subject variable name. The default is \emph{getEctdColName("Subject")}.}

\item{seed}{Random seed. The default is \emph{.deriveFromMasterSeed()}.}

\item{flagName}{Name for omit flag. The default is \emph{getEctdColName("ParOmit")}.}
}
\description{
Creates simulated fixed and between subject parameters for subjects in each replicate
}
