% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTvDefinition.R
\name{createTvDefinition}
\alias{createTvDefinition}
\title{Create "Typical Value" Definition}
\usage{
createTvDefinition(
  drugName,
  drugValue,
  doses,
  times = NULL,
  type = c("Parallel", "Crossover"),
  sequence,
  testCall = TRUE
)
}
\arguments{
\item{drugName}{...}

\item{drugValue}{...}

\item{doses}{...}

\item{times}{...}

\item{type}{...}

\item{sequence}{...}

\item{testCall}{...}
}
\value{
A "tvDefinition" object
}
\description{
Create "Typical Value" Definition for use with "SIMON" typical value
simulation
}
\details{
Constructs "tvDefinition" object
}
\examples{


createTvDefinition("A", 1, 1:3)


}
\author{
Mike K Smith \email{mstoolkit@googlemail.com}
}
\keyword{Typical}
\keyword{Value}
