% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ectdStop.R, R/ectdWarning.R
\name{ectdStop}
\alias{ectdStop}
\alias{ectdWarning}
\title{Error handling functions for the ectd package.}
\usage{
ectdStop(msg, call. = TRUE, domain = NULL, verbose = getEctdVerbose())

ectdWarning(..., call. = TRUE, immediate. = FALSE, domain = NULL)
}
\arguments{
\item{msg}{(Required) Message to be displayed to the user}

\item{call.}{logical, indicating if the call should become part of the
    error message.}

\item{domain}{see \code{\link[base]{gettext}}.  If \code{NA}, messages will
    not be translated.}

\item{verbose}{(Optional) Should more details be given when showing the
error message/warning.  By default, this is controlled by support function
\link{getEctdVerbose}}

\item{...}{zero or more objects which can be coerced to character
    (and which are pasted together with no separator) or a single
    condition object.}

\item{immediate.}{logical, indicating if the call should be output
    immediately, even if \code{\link[base]{getOption}("warn") <= 0}.}
}
\value{
Nothing.
}
\description{
These functions are alias to the standard R functions stop and warning, but
they also do character interpolation. See examples.
}
\section{Functions}{
\itemize{
\item \code{ectdWarning}: Error handling functions for the ectd package.
}}

\examples{

  \dontrun{
     ectdStop("something wrong happened")
  }

}
\seealso{
See \code{\link{warning}} and \code{\link{stop}}.
}
\author{
Rich Pugh
}
\keyword{error}
