\name{MSToolkit-package}
\alias{MSToolkit-package}
\alias{MSToolkit}
\docType{package}
\title{
Generates and analyzes simulated trial data
}
\description{
Originally designed and built for Pfizer by Mango Solutions, the MSToolkit library generates and analyzes simulated trial data.
}
\details{
\tabular{ll}{
Package: \tab MSToolkit\cr
Type: \tab Package\cr
Version: \tab 3.2.0\cr
Date: \tab 2012-09-22\cr
}}

\author{ Mike K Smith \email{mstoolkit@googlemail.com} } 

\keyword{ package }
\seealso{
  Neal Thomas' "DoseResponse" library
}
