% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSToolkit-package.R
\docType{package}
\name{MSToolkit-package}
\alias{MSToolkit}
\alias{MSToolkit-package}
\title{MSToolkit: The MSToolkit Library For Clinical Trial Design}
\description{
Simulate data, analyse this data and assess operating characteristcs of clinical trials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MikeKSmith/MSToolkit}
  \item Report bugs at \url{https://github.com/MikeKSmith/MSToolkit/issues}
}

}
\author{
\strong{Maintainer}: Mike K Smith \email{mstoolkit@gmail.com}

Authors:
\itemize{
  \item Rich Pugh \email{rpugh@mango-solutions.com}
  \item Roman Francois
  \item Francisco Gochez
}

Other contributors:
\itemize{
  \item Jian Li [contributor]
  \item Jinhao Luo [contributor]
  \item Iyer Narayanan Seshan [contributor]
}

}
\keyword{internal}
