% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstoolkitUnitTests.R
\name{mstoolkitUnitTests}
\alias{mstoolkitUnitTests}
\title{Unit/System Tests suite for the MSToolkit package}
\usage{
mstoolkitUnitTests(htmlreport = FALSE, showdetails = TRUE)
}
\arguments{
\item{htmlreport}{(Optional) A logical value.  Print an html report for unit
and system tests?  FALSE by default}

\item{showdetails}{(Optional) A logical value.  Print the details in the
text summary of the unit tests?  TRUE by default}
}
\value{
The function does not return anything. It writes two files:
\code{reportUT.html} for the unit tests and \code{reportST.html} for the
system tests.
}
\description{
This function enables the user of the MSToolkit package to run the unit
tests of the package.
}
\details{
The unit tests are based on the RUnit framework.
}
\examples{

  \dontrun{
    mstoolkitUnitTests( htmlhelp = TRUE )

    # unit tests
    browseURL("reportUT.html")

    # system tests
    browseURL("reportST.html")

  }

}
\references{
Matthias Burger, Klaus Juenemann and Thomas Koenig (2007).
RUnit: R Unit test framework. R package version 0.4.17.
https://sourceforge.net/projects/runit/

RUnit and R CMD check
http://wiki.r-project.org/rwiki/doku.php?id=developers:runit
}
\seealso{
\code{\link[RUnit]{runTestSuite}}
}
\author{
Mike K Smith \email{mstoolkit@googlemail.com}
}
\keyword{error}
