% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaManagement.R
\name{setEctdVerbose}
\alias{setEctdVerbose}
\alias{setEctdLogFile}
\alias{setEctdDateFormat}
\alias{getEctdVerbose}
\alias{getEctdLogFile}
\alias{getEctdDateFormat}
\title{MSToolkit package options}
\usage{
setEctdVerbose(verbose)
}
\arguments{
\item{verbose}{(Required) A logical value.  If set to TRUE, messages are
sent to the logfile during the process of generating the data and analyzing
it. Set to \code{TRUE} when attaching the package.}
}
\value{
The function (invisibly) returns the previous value of the
arguments.
}
\description{
Options used by the MSToolkit package to control the logfile, the amount of
messages that are written in the logfile, and the format of the date.
}
\details{
The three function write and read information from the (not exported)
environment \code{.ectdEnv}.  These settings are mainly used by the (not
exported) \code{.log} function.
}
\examples{
\dontrun{
  oldverb <- setEctdVerbose( TRUE )
  olddf   <- setEctdDateFormat("\%Y")
  oldlf   <- setEctdLogFile("mstoolkit.log")

  for( i in 1:100 ) {
    MSToolkit:::.log( paste("some message:", i) )
  }
  file.show( getEctdLogFile() )

  setEctdVerbose   (oldverb)
  setEctdDateFormat(olddf  )
  setEctdLogFile   (oldlf  )

}

}
\seealso{
\code{\link{options}} provides a similar mechanism for R options.
}
\keyword{IO}
