% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validNames.R
\name{validNames}
\alias{validNames}
\title{Check if a character string is a valid R variable name}
\usage{
validNames(...)
}
\arguments{
\item{\dots}{(Required) A collection of character vectors to be checked}
}
\value{
None. The function is simply generates an error if an invalid name
is passed to it.
}
\description{
Checks if a character string is a valid R variable name using regular
expressions.
}
\details{
A valid R name starts with a letter or a dot followed by a non-numerical
character and contains only letters, numbers and dots.
}
\examples{

  # correct name
  validNames("DOSE")
  \dontrun{
    # wrong name
    validNames(".334fsedqw")
  }

}
\author{
Romain Francois
}
\keyword{error}
